+++ src/main/java/com/skriptvariables/util/VariableStore.java
@@ -1,0 +1,8 @@
+package com.skriptvariables.util;
+
+public interface VariableStore {
+
+    Object get(String name);
+
+    void set(String name, Object value);
+}